requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
Django==4.1.6
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.4
openai==0.26.5
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.0.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
zope.interface==5.5.2
